Safari-609.1.11.1/LayoutTests/svg/dom/SVGPathSegList-removeItem-expected.txt

This is a test of the SVGPathSegList::removeItem() API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial 'pathSegList' value of path1
PASS path1.pathSegList.numberOfItems is 5
PASS path1.pathSegList.getItem(0).toString() is "[object SVGPathSegMovetoAbs]"
PASS path1.pathSegList.getItem(0).x is 0
PASS path1.pathSegList.getItem(0).y is 0
PASS path1.pathSegList.getItem(1).toString() is "[object SVGPathSegLinetoAbs]"
PASS path1.pathSegList.getItem(1).x is 100
PASS path1.pathSegList.getItem(1).y is 0
PASS path1.pathSegList.getItem(2).toString() is "[object SVGPathSegLinetoAbs]"
PASS path1.pathSegList.getItem(2).x is 100
PASS path1.pathSegList.getItem(2).y is 100
PASS path1.pathSegList.getItem(3).toString() is "[object SVGPathSegClosePath]"
PASS path1.pathSegList.getItem(4).toString() is "[object SVGPathSegLinetoAbs]"
PASS path1.pathSegList.getItem(4).x is 0
PASS path1.pathSegList.getItem(4).y is 100

Check initial 'pathSegList' value of path2
PASS path2.pathSegList.numberOfItems is 5
PASS path2.pathSegList.getItem(0).toString() is "[object SVGPathSegMovetoAbs]"
PASS path2.pathSegList.getItem(0).x is 0
PASS path2.pathSegList.getItem(0).y is 0
PASS path2.pathSegList.getItem(1).toString() is "[object SVGPathSegLinetoHorizontalRel]"
PASS path2.pathSegList.getItem(1).x is 100
PASS path2.pathSegList.getItem(2).toString() is "[object SVGPathSegClosePath]"
PASS path2.pathSegList.getItem(3).toString() is "[object SVGPathSegLinetoVerticalRel]"
PASS path2.pathSegList.getItem(3).y is 100
PASS path2.pathSegList.getItem(4).toString() is "[object SVGPathSegLinetoHorizontalRel]"
PASS path2.pathSegList.getItem(4).x is -100

Remove fourth item of path1
PASS path1.pathSegList.removeItem(3).toString() is "[object SVGPathSegClosePath]"

Remove third item of path2
PASS path2.pathSegList.removeItem(2).toString() is "[object SVGPathSegClosePath]"
PASS successfullyParsed is true

TEST COMPLETE
